# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Discord Music API — Manages music playback via Lavalink v4
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
  /** @nullable */
  details?: string | null;
}

export interface TrackInfo {
  identifier: string;
  title: string;
  author: string;
  /** Duration in milliseconds */
  duration: number;
  /** @nullable */
  uri?: string | null;
  /** @nullable */
  artworkUrl?: string | null;
  sourceName: string;
  isStream: boolean;
  isSeekable: boolean;
}

export interface Track {
  encoded: string;
  info: TrackInfo;
  /**
     * Discord user ID who requested this track
     * @nullable
     */
  requester?: string | null;
}

export type SearchResultLoadType = typeof SearchResultLoadType[keyof typeof SearchResultLoadType];


export const SearchResultLoadType = {
  search: 'search',
  track: 'track',
  playlist: 'playlist',
  empty: 'empty',
  error: 'error',
} as const;

export interface SearchResult {
  loadType: SearchResultLoadType;
  tracks: Track[];
  /** @nullable */
  playlistName?: string | null;
  query: string;
  source: string;
  cached: boolean;
}

export type PlayInputSource = typeof PlayInputSource[keyof typeof PlayInputSource];


export const PlayInputSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
  spotify: 'spotify',
  ytsearch: 'ytsearch',
  scsearch: 'scsearch',
} as const;

export interface PlayInput {
  /** Search query, URL, or encoded track */
  query: string;
  /** Discord voice channel ID to join */
  channelId: string;
  /**
     * Discord user ID of the requester
     * @nullable
     */
  requesterId?: string | null;
  source?: PlayInputSource;
  /** Add to queue if something is already playing */
  addToQueue?: boolean;
}

export type PlayResponseStatus = typeof PlayResponseStatus[keyof typeof PlayResponseStatus];


export const PlayResponseStatus = {
  playing: 'playing',
  queued: 'queued',
} as const;

export interface PlayResponse {
  status: PlayResponseStatus;
  track: Track;
  /**
     * Position in queue (1-indexed, null if now playing)
     * @nullable
     */
  position?: number | null;
  queueSize: number;
}

export interface PlaylistInput {
  /** Playlist URL (YouTube, Spotify, SoundCloud) */
  url: string;
  /** Discord voice channel ID to join */
  channelId: string;
  /**
     * Discord user ID of the requester
     * @nullable
     */
  requesterId?: string | null;
  /** Shuffle the playlist before adding */
  shuffle?: boolean;
}

export type PlaylistResponseStatus = typeof PlaylistResponseStatus[keyof typeof PlaylistResponseStatus];


export const PlaylistResponseStatus = {
  playing: 'playing',
  queued: 'queued',
} as const;

export interface PlaylistResponse {
  status: PlaylistResponseStatus;
  playlistName: string;
  tracksLoaded: number;
  queueSize: number;
  firstTrack: Track;
}

export interface PlayerActionResponse {
  success: boolean;
  message: string;
  guildId: string;
}

export interface SkipResponse {
  success: boolean;
  skippedTrack: Track;
  /** @nullable */
  nextTrack?: Track | null;
  queueSize: number;
}

export interface VolumeInput {
  /**
     * Volume level (0-1000, 100 = normal)
     * @minimum 0
     * @maximum 1000
     */
  volume: number;
}

/**
 * Loop mode
 */
export type LoopInputMode = typeof LoopInputMode[keyof typeof LoopInputMode];


export const LoopInputMode = {
  none: 'none',
  track: 'track',
  queue: 'queue',
} as const;

export interface LoopInput {
  /** Loop mode */
  mode: LoopInputMode;
}

export type NowPlayingResponseLoopMode = typeof NowPlayingResponseLoopMode[keyof typeof NowPlayingResponseLoopMode];


export const NowPlayingResponseLoopMode = {
  none: 'none',
  track: 'track',
  queue: 'queue',
} as const;

export interface NowPlayingResponse {
  track: Track;
  /** Current position in milliseconds */
  position: number;
  paused: boolean;
  volume: number;
  loopMode: NowPlayingResponseLoopMode;
  queueSize: number;
}

export type PlayerStatusLoopMode = typeof PlayerStatusLoopMode[keyof typeof PlayerStatusLoopMode];


export const PlayerStatusLoopMode = {
  none: 'none',
  track: 'track',
  queue: 'queue',
} as const;

export interface PlayerStatus {
  guildId: string;
  connected: boolean;
  /** @nullable */
  channelId?: string | null;
  playing: boolean;
  paused: boolean;
  volume: number;
  loopMode: PlayerStatusLoopMode;
  shuffleEnabled: boolean;
  queueSize: number;
  /** @nullable */
  currentTrack?: Track | null;
  /** Current position in milliseconds */
  position: number;
  /** Node ping in milliseconds */
  ping: number;
}

export interface QueueResponse {
  guildId: string;
  tracks: Track[];
  total: number;
  /** @nullable */
  currentTrack?: Track | null;
}

export interface LavalinkNodeStats {
  name: string;
  url: string;
  connected: boolean;
  players: number;
  playingPlayers: number;
  uptime: number;
  memoryUsed: number;
  memoryFree: number;
  cpuCores: number;
  cpuLoad: number;
  ping: number;
}

export interface NodesStatusResponse {
  totalNodes: number;
  connectedNodes: number;
  totalPlayers: number;
  nodes: LavalinkNodeStats[];
}

export type SearchTracksParams = {
/**
 * Search query or direct URL (YouTube, Spotify, SoundCloud)
 */
query: string;
/**
 * Search source
 */
source?: SearchTracksSource;
/**
 * Maximum number of results
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type SearchTracksSource = typeof SearchTracksSource[keyof typeof SearchTracksSource];


export const SearchTracksSource = {
  youtube: 'youtube',
  soundcloud: 'soundcloud',
  spotify: 'spotify',
  ytsearch: 'ytsearch',
  scsearch: 'scsearch',
} as const;