# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Discord Music API — Manages music playback via Lavalink v4
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /music/guilds/{guildId}/shuffle:
    post:
      operationId: shuffleQueue
      tags: [queue]
      summary: Shuffle queue
      description: Shuffle the current queue randomly
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Queue shuffled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerActionResponse"
        "400":
          description: Queue empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /music/guilds/{guildId}/queue:
    get:
      operationId: getQueue
      tags: [queue]
      summary: Get the queue
      description: Retrieve all tracks in the current playback queue
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Queue contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No active player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: clearQueue
      tags: [queue]
      summary: Clear queue
      description: Remove all tracks from the queue (keeps current track playing)
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Queue cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerActionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No active player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    GuildId:
      name: guildId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9]{17,20}$"
      description: Discord Guild (server) ID
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
    TrackInfo:
      type: object
      properties:
        identifier:
          type: string
        title:
          type: string
        author:
          type: string
        duration:
          type: integer
          description: Duration in milliseconds
        uri:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        sourceName:
          type: string
        isStream:
          type: boolean
        isSeekable:
          type: boolean
      required:
        - identifier
        - title
        - author
        - duration
        - sourceName
        - isStream
        - isSeekable
    Track:
      type: object
      properties:
        encoded:
          type: string
        info:
          $ref: "#/components/schemas/TrackInfo"
        requester:
          type: ["string", "null"]
          description: Discord user ID who requested this track
      required:
        - encoded
        - info
...
    PlayerActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        guildId:
          type: string
      required:
        - success
        - message
        - guildId
...
    QueueResponse:
      type: object
      properties:
        guildId:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
        currentTrack:
          type: ["object", "null"]
          $ref: "#/components/schemas/Track"
      required:
        - guildId
        - tracks
        - total
...